+++ packages/core/src/index.ts
@@ -9,10 +9,20 @@ export { OnePasswordError, resolveOnePasswordReference } from "./onepassword.js"
 export { getAuthStatus, resolveCredentials, signIn, SignInError } from "./auth.js";
 export type { AuthInvalidReason, AuthStatusResult, SignInErrorCode } from "./auth.js";
 
+export { AuthRevokedError, TtctlError } from "./auth/errors.js";
+
 export { deleteAuthToken, loadAuthToken, resolveAuthTokenPath, saveAuthToken } from "./authToken.js";
 export type { ResolveAuthTokenPathOptions } from "./authToken.js";
 
-export { Cf403Error, IMPERSONATE_PROFILE, callSurface, impersonatedTransport, stockTransport } from "./transport.js";
+export {
+  Cf403Error,
+  Cf403PersistentError,
+  IMPERSONATE_PROFILE,
+  SchedulerBearerExpired,
+  callSurface,
+  impersonatedTransport,
+  stockTransport,
+} from "./transport.js";
 export type { TransportRequest, TransportResponse } from "./transport.js";
 
 export { SURFACES_REQUIRING_IMPERSONATION, SURFACE_ENDPOINTS } from "./types.js";